fastapi-app/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
bcrypt==3.2.2
bidict==0.22.0
black==22.3.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.0
fakeredis==1.7.1
fastapi==0.75.2
fastapi-mail==1.0.8
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.6.8
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
pyodbc==4.0.32
pyparsing==3.0.8
python-dotenv==0.20.0
python-engineio==4.3.2
python-multipart==0.0.5
python-socketio==5.6.0
PyYAML==6.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
wrapt==1.14.0
